Create 1-10-22_IPS.java
Word Vertification [1047_IPS-4]
A word is called as a good word if all the letters of the word are distinct. That is, all the letters of the word are different from each other letter. Else, the word is called as a bad word.

Write Java code to check if the given word is good or bad.:

 

 e.g. START, GOOD, BETTER are bad: WRONG is good!

Input format:

A word

Output format:

Print ‘Good’ if all letters of the word are distinct and print ‘Bad’ otherwise